Crate faasta_interface

Structs

FunctionInfo

Represents a published function

FunctionMetricsResponse

Function metrics information

FunctionServiceClient

The client stub that makes RPC calls to the server. All request methods return Futures.

FunctionServiceImpl

Implementation of the FunctionService

Metrics

Overall metrics information

ServeFunctionService

A serving function to use with [::tarpc::server::InFlightRequest::execute].

Enums

FunctionError

FunctionServiceRequest

The request sent over the wire from the client to the server.

FunctionServiceResponse

The response sent over the wire from the server to the client.

Constants

MAX_WASM_SIZE

Traits

FunctionService

Service interface for managing functions

FunctionServiceStub

The stub trait for service FunctionService.

Functions

create_service_with_github_auth

Helper function to create a service implementation with GitHub auth

Type Aliases

AuthValidatorFn

Type alias for the auth validator function type

FunctionResult